pyproject.toml
[project]
name = "information-retrieval-with-diversity"
version = "0.1.0"
description = "Add a short description here"
authors = [
    { name = "tubo213", email = "yuuki.32040808@gmail.com" }
]
dependencies = [
    "numpy>=1.24.4",
    "matplotlib>=3.7.2",
    "tqdm>=4.65.0",
    "torch>=2.0.1",
    "joblib>=1.3.1",
    "pyyaml>=6.0.1",
    "types-PyYAML>=6.0.12.11",
    "hydra-core>=1.3.2",
    "pytorch_pfn_extras>=0.7.1",
    "scipy>=1.10.1",
    "pandas>=2.0.3",
    "pytorch-lightning>=2.0.6",
    "polars>=0.18.9",
    "scikit-learn>=1.3.0",
    "gensim>=4.3.1",
    "faiss-cpu>=1.7.4",
    "streamlit>=1.25.0",
]
readme = "README.md"
requires-python = ">= 3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=23.7.0",
    "flake8>=5.0.4",
    "isort>=5.12.0",
    "mypy>=1.4.1",
    "jupyter>=1.0.0",
]

[tool.black]
line-length = 99

[tool.flake8]
max-line-length = 99
max-complexity = 20
ignore = [
    "W293",
    "W503",
    "W504",
    "E402",
    "F401",
    "F403",
    "E203"
]


[tool.hatch.metadata]
allow-direct-references = true
